Yield at the end of each kthread loop to prevent RCU from detecting a stall
If the RCU detects a stall it collects a full stack trace and writes it to the kernel message buffer. This takes a lot of time and is undesireable. By yielding and setting isolcpus we do still get all the CPU time we need while not being interrupted by the RCU.