Add DMA channel debugging
The BCM2835 has two different types of DMA channels. Channel 0 to 6 are fully featured DMA channels. Channel 7 to 14 have a reduced feature set, rendering them useless for adamtx.
Unfortunately 3 of the 6 fully featured DMA channels are always occupied by the ARM firmware. This leaves only 3 of the for the OS. Both Archlinuxarm and Raspbian are configured to
use one more channel for the sd host controller and - on the newer hardware versions - another two for SDIO. Thus there may be no channels available.